requirements.txt
asn1crypto==0.24.0
astroid==1.6.2
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkginfo==1.4.2
pycparser==2.18
pylint==1.8.3
python-dateutil==2.7.0
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
tqdm==4.19.8
twine==1.11.0
urllib3==1.22
wrapt==1.10.11
